using shorts for vertical positions is not a good idea. Fixes rendering problems on long pages.
svn path=/trunk/kdelibs/; revision=101147

+++ khtml/rendering/bidi.h
@@ -70,7 +70,7 @@ namespace khtml {
 	    }
 	}
 
-	short vertical;
+	int vertical;
 	short baseline;
 	short height;
 	int width;
